+++ kitty/history.c
@@ -59,11 +59,19 @@ alloc_pagerhist(unsigned int pagerhist_sz) {
     PagerHistoryBuf *ph;
     if (!pagerhist_sz) return NULL;
     ph = PyMem_Calloc(1, sizeof(PagerHistoryBuf));
+    if (!ph) return NULL;
     ph->maxsz = pagerhist_sz / sizeof(Py_UCS4);
     ph->bufsize = 1024*1024 / sizeof(Py_UCS4);
     ph->buffer = PyMem_RawMalloc(1024*1024);
     if (!ph->buffer) { PyMem_Free(ph); return NULL; }
     return ph;
+}
+
+static inline void
+free_pagerhist(HistoryBuf *self) {
+    if (self->pagerhist) PyMem_Free(self->pagerhist->buffer);
+    PyMem_Free(self->pagerhist);
+    self->pagerhist = NULL;
 }
 
 static inline bool
@@ -75,6 +83,14 @@ pagerhist_extend(PagerHistoryBuf *ph, size_t minsz) {
     ph->buffer = newbuf;
     ph->bufsize = newsz;
     return true;
+}
+
+static inline void
+pagerhist_clear(HistoryBuf *self) {
+    if (!self->pagerhist || !self->pagerhist->maxsz) return;
+    index_type pagerhist_sz = self->pagerhist->maxsz  * sizeof(Py_UCS4);
+    free_pagerhist(self);
+    self->pagerhist = alloc_pagerhist(pagerhist_sz);
 }
 
 static PyObject *
@@ -112,8 +128,7 @@ dealloc(HistoryBuf* self) {
         PyMem_Free(self->segments[i].line_attrs);
     }
     PyMem_Free(self->segments);
-    if (self->pagerhist) PyMem_Free(self->pagerhist->buffer);
-    PyMem_Free(self->pagerhist);
+    free_pagerhist(self);
     Py_TYPE(self)->tp_free((PyObject*)self);
 }
 
@@ -152,8 +167,9 @@ historybuf_mark_line_dirty(HistoryBuf *self, index_type y) {
     *p |= TEXT_DIRTY_MASK;
 }
 
-inline void
+void
 historybuf_clear(HistoryBuf *self) {
+    pagerhist_clear(self);
     self->count = 0;
     self->start_of_data = 0;
 }